Added logic for handling SSL certificate
This change allows the charm to read the ssl_certificate if its provided
by the charms config. The Helper code handles attempting to write this
to the prescribed file and then updates the system's certificates. The
charm code will block if theres a problem updating the certificates,
when one is provided, but otherwise will allow checks to be run without
a certificate. This can be change fairly easily in the future, as well
as where the certificate can be read from, which should actually be a
from a relation.